MAINT: matrix.approximation.positive_semidefinite.Reimer: use float128 instead of float64 internally

+++ matrix/approximation/positive_semidefinite/Reimer.py
@@ -280,7 +280,7 @@ def _decomposition(
             gamma = gamma.real
 
         # data type to use for calculations
-        DTYPE = np.float64
+        DTYPE = np.float128
         d_eps = np.finfo(DTYPE).eps
 
         # check diag values
